# openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/resources": {
      "get": {
        "operationId": "ResourcesV1Controller_listResources",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public resource list"
          }
        },
        "tags": [
          "v1-resources"
        ]
      },
      "post": {
        "operationId": "ResourcesV1WriteController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "resource_type",
                  "version"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "resource_type": {
                    "type": "string",
                    "enum": [
                      "upload",
                      "external"
                    ]
                  },
                  "resource_kind": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "description": "Legacy Markdown projection; kept for compatibility."
                  },
                  "content_json": {
                    "type": "object",
                    "description": "Canonical Tiptap/ProseMirror document. Submitted as a JSON string in multipart requests."
                  },
                  "preview_draft_id": {
                    "type": "string"
                  },
                  "schematic_code": {
                    "type": "string"
                  },
                  "external_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submitted to the existing resource moderation lifecycle."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
    "/v1/resources/{id}/manifest": {
      "get": {
        "operationId": "ResourcesV1Controller_getManifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Launcher and in-game resource manifest"
          }
        },
        "tags": [
          "v1-resources"
        ]
      }
    },
    "/v1/resources/{id}/preview": {
      "get": {
        "operationId": "ResourcesV1Controller_getPreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "v1-resources"
        ]
      }
    },
    "/v1/resources/{resourceId}/versions/{versionId}/files/{fileId}/download": {
      "get": {
        "operationId": "ResourcesV1Controller_downloadFile",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "v1-resources"
        ]
      }
    },
    "/v1/resources/{id}": {
      "get": {
        "operationId": "ResourcesV1Controller_getResource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource detail"
          }
        },
        "tags": [
          "v1-resources"
        ]
      }
    },
...
    "/v1/threads": {
      "get": {
        "operationId": "ThreadsV1Controller_listThreads",
        "parameters": [
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads"
          }
        },
        "tags": [
          "v1-threads"
        ]
      },
      "post": {
        "operationId": "ThreadWriteV1Controller_createThread",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thread created. It can be pending moderation."
          }
        },
        "tags": [
          "v1-thread-writes"
        ]
      }
    },
...
    "/v1/threads/{id}/like": {
      "put": {
        "operationId": "ThreadInteractionsV1Controller_like",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread is liked; safe to repeat."
          }
        },
        "tags": [
          "v1-thread-interactions"
        ]
      },
      "delete": {
        "operationId": "ThreadInteractionsV1Controller_unlike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread is not liked; safe to repeat."
          }
        },
        "tags": [
          "v1-thread-interactions"
        ]
      }
    },
    "/v1/threads/{id}/bookmark": {
      "put": {
        "operationId": "ThreadInteractionsV1Controller_bookmark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread is bookmarked; safe to repeat."
          }
        },
        "tags": [
          "v1-thread-interactions"
        ]
      },
      "delete": {
        "operationId": "ThreadInteractionsV1Controller_removeBookmark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread is not bookmarked; safe to repeat."
          }
        },
        "tags": [
          "v1-thread-interactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "MDTBBS Public Client API",
    "description": "Stable V1 contract for web, official applications and third-party public clients. Use /api/v1/capabilities for feature discovery. External server integrations use the separate /api/external/v1 API key boundary. JSON V1 endpoints return the { data, meta } / { error, meta } envelope unless explicitly documented as a raw file or image response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "Same-origin MindFourm API"
    }
  ],
  "components": {
    "securitySchemes": {
      "MindAuthBearer": {
        "scheme": "bearer",
        "bearerFormat": "MindAuth access token",
        "type": "http"
      }
    },
    "schemas": {
...
      "TiptapMarkDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "bold",
              "italic",
              "strike",
              "underline",
              "code",
              "link"
            ]
          },
          "attrs": {
            "type": "object",
            "additionalProperties": true,
            "description": "Link marks accept href and optional title."
          }
        },
        "required": [
          "type"
        ]
      },
      "TiptapNodeDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "doc",
              "paragraph",
              "heading",
              "blockquote",
              "bulletList",
              "orderedList",
              "listItem",
              "codeBlock",
              "horizontalRule",
              "table",
              "tableRow",
              "tableHeader",
              "tableCell",
              "text",
              "hardBreak",
              "image"
            ]
          },
          "attrs": {
            "type": "object",
            "additionalProperties": true,
            "description": "Only attributes defined by the named node type are accepted."
          },
          "text": {
            "type": "string",
            "description": "Text for a text node."
          },
          "content": {
            "description": "Child nodes in the allowlisted Tiptap schema.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TiptapNodeDto"
            }
          },
          "marks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TiptapMarkDto"
            }
          }
        },
        "required": [
          "type"
        ]
      },
      "TiptapDocumentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "doc"
            ]
          },
          "content": {
            "description": "Root blocks; recursive node schema is bounded and validated by MindFourm.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TiptapNodeDto"
            }
          }
        },
        "required": [
          "type",
          "content"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Legacy Markdown projection. May be omitted when content_json is supplied."
          },
          "content_json": {
            "description": "Tiptap / ProseMirror JSON source. The server validates against its allowlisted schema.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TiptapDocumentDto"
              }
            ]
          }
        }
      },
...
    }
  }
}